Alarms Page with history tab
