+++ src/sass/app.scss
@@ -90,6 +90,10 @@ html {
   height: 100%;
 }
 
+.full-width {
+  width: 100%;
+}
+
 .half-height {
   height: 50%;
 }
